Route middleware added

+++ routes/web.php
@@ -15,35 +15,35 @@ use App\Models\Listings;
 | contains the "web" middleware group. Now create something great!
 |
 */
-
+//show all listings
 Route::get('/listings', [ListingController::class, 'index']);
+//show listings form
+Route::get('/listings/create',[ListingController::class,'create'])->middleware('auth');
+//store listings data
+Route::post('/listings', [ListingController::class,'store'])->middleware('auth');
 
-Route::get('/listings/create',[ListingController::class,'create']);
-
-Route::post('/listings', [ListingController::class,'store']);
-
-
+//show single listing
 Route::get('/listings/{listing}',[ListingController::class, 'show']);
 
 
 
 //show edit form
-Route::get('/listings/{listing}/edit', [ListingController::class,'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class,'edit'])->middleware('auth');
 //update listing
-Route::put('listings/{listing}',[ListingController::class,'update']);
+Route::put('listings/{listing}',[ListingController::class,'update'])->middleware('auth');
 //Delete listing
-Route::delete('/listings/{listing}',[ListingController::class,'destroy']);
+Route::delete('/listings/{listing}',[ListingController::class,'destroy'])->middleware('auth');
 
 //show register form
-Route::get('/register',[UserController::class,'create']);
+Route::get('/register',[UserController::class,'create'])->middleware('guest');
 //create new user
 Route::post('/users/create',[UserController::class,'store']);
 
 //Log user out
-Route::post('/logout',[UserController::class,'logout']);
+Route::post('/logout',[UserController::class,'logout'])->middleware('auth');
 
 //Show login form
-Route::get('/login',[UserController::class,'login']);
+Route::get('/login',[UserController::class,'login'])->name('login')->middleware('guest');
 //Login user
 Route::post('/users/authenticate',[UserController::class,'authenticate']);
